requirements.txt
astroid==3.0.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
dill==0.3.7
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
oauthlib==3.2.2
platformdirs==3.11.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pylint==3.0.1
pyparsing==3.1.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
